Added ability to load existing note, but it flow have bug with null (WIP).

+++ app/src/main/java/com/dvinc/notepad/data/database/dao/NoteDao.kt
@@ -25,7 +25,7 @@ interface NoteDao {
     fun deleteNote(noteId: Int)
 
     @Query("SELECT * FROM Notes WHERE id = :noteId")
-    fun getNoteById(noteId: Int): NoteEntity
+    fun getNoteById(noteId: Long): NoteEntity
 
     @Update
     fun updateNote(note: NoteEntity)
